requirements.txt
babel==2.10.3
django_cors_headers==3.13.0
django==4.1.3
djangorestframework==3.14.0
einops==0.6.0
gdown==4.5.3
keras==2.10.0
ninja==1.10.2.4
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.2.0
psycopg2==2.9.5
pyclipper==1.3.0
pycodestyle==2.9.1
pydantic==1.10.2
python-bidi==0.4.2
PyYAML==6.0
scikit-image==0.19.3
scipy==1.7.3
Shapely==1.8.5
tensorflow==2.10.0
torch==1.12.1
torchvision==0.13.1
Unidecode==1.3.6
